cleaning test_ file and moving central to harness
